[BOLT] Add minimal LoongArch 64-bit support
Just enough features are implemented to process a simple "hello
world" executable and produce something that still runs (including
libc calls). This was mainly a matter of implementing support for
various relocations. Currently, the following are handled:

- R_LARCH_B26
- R_LARCH_PCALA_LO12
- R_LARCH_PCALA_HI20
- R_LARCH_GOT_PC_LO12
- R_LARCH_GOT_PC_HI20
- R_LARCH_GOT64_PC_LO20
- R_LARCH_GOT64_PC_HI12
- R_LARCH_NONE

Tests are provided for all supported relocations. Refer to D145687,
an ELF file produced by GCC had to be used in order to test the
correct handling of PLT entries, although the YAML representation
is quite large.

Change-Id: Ic9f8af89aab4c0e954f8f4e37b6521c2f945d5ef