[BOLT][LoongArch] Implement pass to expand bl calls and b tail-calls to pcaddu18i+jirl
The mainly purpose of this patch is to fix the Branch26PCRel out of
range error occurs when linking by JITLink.

The simplest case is when the bl instruction in .text section wants to
call .plt section and there is also a big .bss section. The new .text
section created by BOLT is placed behind .bss, so it maybe very far from
.plt section. In this case, bl instr's 28-bit offset will be not enough
and an out of range error occurs.

Expand all bl calls and b tail-calls to pcaddu18i+jirl used in medium
code model is a temporary method. But the error can also occurs when
pcaddu18i+jirl also cannot reach the call destination. In this case,
extreme code model call will needed. Leaving this problem to be solved
later and waiting for further consideration.

Redmine: http://sysdev.loongson.cn/issues/33609
Change-Id: I762c942f8723456f6b7392adeca60051b752882f