[LoongArch][GlobalISel] Legalize aligned/unaligned [ext]load and store opcodes
Legalize power of 2 bytes ~ 2xGRLen types with and without
alignment loads and stores, and sext/zext loads.

Set Alignment to 8 when `ual` is allowed. Otherwise, when
legalizer try to lower the unaligned load/store, allowsMemoryAccess
hook called by the lower code will say the unaligned access is
allowed. So the legalizer gives up and leads to Legalization failure.