[LoongArch] Expand pseudos later and attach relax relocs when relax enabled
If linker relaxation enabled, pseudos that can be relaxed should be
expanded later to avoid being separated by instruction scheduling.
This commit delays expanding them (except for `tls_le`, because it
can be scheduled) in LoongArchExpandPseudo pass.

A new mask target-flag is added to attach relax relocs to the
relaxable code sequence.