[clang-repl][LoongArch] Make feature relax in jit consistent with driver
This commit enables feature relax for loongarch64 when entering
JIT so that it can be consistent with the clang driver.

Without this commit, JIT cannot get feature relax after calling
`orc::JITTargetMachineBuilder::detectHost()`, thus the return
result of `LoongArchELFObjectWriter::needsRelocateWithSymbol()`
will be `false` which means that relocs will use the section
symbol plus offset representation. But linker relaxation requests
must use the real symbol offset in the symtab, because it can
only update the symtab after relaxing but cannot update the
addend for any edge. So error will occur because of the wrong
target address (symbol address) of edges getting by jitlink.

Note: In fact, riscv also has this issue. But the return result
of `RISCVELFObjectWriter::needsRelocateWithSymbol()` is always
`true`. So this will never happen.